Extract node-prefix-path
