+++ src/file-system/index.js
@@ -1,6 +1,7 @@
 'use babel'
 
-export { default as getConfigPath } from './get-config-path'
-export { default as cdToProjectRoot } from './root-path'
-export { getIgnoreFile } from './ignore-file'
+export { default as getConfigPath } from './config-path'
 export * from './fs-utils'
+export { getIgnoreFile } from './ignore-file'
+export { default as getNodePrefixPath } from './node-prefix-path'
+export { default as cdToProjectRoot } from './root-path'
